+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Breaking
+- Test no longer pass if any plugin failed to load
+
 ### Added
 - `DirectoryTestConfig.overrideDocumentationOutput` allows to overwrite documentation similar to `overrideParserOutput`